+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # v1.x.x
 
-*
+* Add Predict methods to Callback to allow callback to be call during prediction phase.
+* Add Predict methods to Experiment to more easily make predictions with an experiment.
+* Add a progress bar callback during predictions of a model.
+
 
 # v1.5